file changes and fixed HTML tags bug with MySQL
