+++ doc/LICENSE.txt
