+++ doc/cheatsheet_2015-11-17.txt
