--- sdk/libs/soil/projects/VC7.1/SOIL.sln
@@ -1,21 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 8.00
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SOIL", "SOIL.vcproj", "{35D9B7E3-EE73-4C06-9B98-FCB7F7644C99}"
-	ProjectSection(ProjectDependencies) = postProject
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfiguration) = preSolution
-		Debug = Debug
-		Release = Release
-	EndGlobalSection
-	GlobalSection(ProjectConfiguration) = postSolution
-		{35D9B7E3-EE73-4C06-9B98-FCB7F7644C99}.Debug.ActiveCfg = Debug|Win32
-		{35D9B7E3-EE73-4C06-9B98-FCB7F7644C99}.Debug.Build.0 = Debug|Win32
-		{35D9B7E3-EE73-4C06-9B98-FCB7F7644C99}.Release.ActiveCfg = Release|Win32
-		{35D9B7E3-EE73-4C06-9B98-FCB7F7644C99}.Release.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-	EndGlobalSection
-	GlobalSection(ExtensibilityAddIns) = postSolution
-	EndGlobalSection
-EndGlobal
