--- sdk/libs/soil/projects/codeblocks/SOIL.cbp
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="Simple OpenGL Image Library" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="lib-Release">
-				<Option output="..\..\lib\libSOIL" prefix_auto="1" extension_auto="1" />
-				<Option working_dir="" />
-				<Option object_output="obj\Release\" />
-				<Option type="2" />
-				<Option compiler="gcc" />
-				<Option createDefFile="1" />
-				<Compiler>
-					<Add option="-Os" />
-					<Add option="-O2" />
-				</Compiler>
-				<Linker>
-					<Add option="-s" />
-				</Linker>
-			</Target>
-			<Target title="test-Debug">
-				<Option output="..\..\testSOIL" prefix_auto="1" extension_auto="1" />
-				<Option working_dir="..\..\" />
-				<Option object_output="obj\" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-O" />
-					<Add option="-pg" />
-					<Add option="-g" />
-				</Compiler>
-				<Linker>
-					<Add option="-pg -lgmon" />
-					<Add library="opengl32" />
-					<Add library="gdi32" />
-				</Linker>
-			</Target>
-			<Target title="test-Release">
-				<Option output="..\..\testSOIL" prefix_auto="1" extension_auto="1" />
-				<Option working_dir="..\..\" />
-				<Option object_output="obj\" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-Os" />
-					<Add option="-O2" />
-				</Compiler>
-				<Linker>
-					<Add option="-s" />
-					<Add library="opengl32" />
-					<Add library="gdi32" />
-				</Linker>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="-Wall" />
-		</Compiler>
-		<Unit filename="..\..\src\SOIL.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="..\..\src\SOIL.h" />
-		<Unit filename="..\..\src\image_DXT.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="..\..\src\image_DXT.h" />
-		<Unit filename="..\..\src\image_helper.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="..\..\src\image_helper.h" />
-		<Unit filename="..\..\src\stb_image_aug.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="..\..\src\stb_image_aug.h" />
-		<Unit filename="..\..\src\stbi_DDS_aug.h" />
-		<Unit filename="..\..\src\stbi_DDS_aug_c.h" />
-		<Unit filename="..\..\src\test_SOIL.cpp">
-			<Option target="test-Debug" />
-			<Option target="test-Release" />
-		</Unit>
-		<Extensions>
-			<code_completion />
-			<envvars />
-			<debugger />
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
